TicketZen/requirements.txt
Flask==3.0.3
Jinja2==3.1.4
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0
PyYAML==6.0.2
rapidfuzz==3.9.6
python-dateutil==2.9.0.post0
